ENH: greatly improve speed of np.pad for most modes
The previous method, using a vector function and
np.apply_along_axis, scaled poorly with dimensionality
(particularly evident on arrays with rank > 3).

The new method iteratively expands the array along each
axis, using helper functions which calculate the entire
extension for a particular axis and then concatenate it
with the input array.

Execution time decreases are considerable at low
dimensionality, and scaling with dimensionality is vastly
improved.

The old method is retained for the user-valued padding
function, as this case is not easily extended into the
new framework.